fix: extract shared test fixtures to conftest.py for LOC compliance
Fix CI blocker: test files exceeded 400-line CI block threshold.

## Problem
- test_dispatch_queue_operations.py: 438 lines (exceeded by 38)
- test_dispatch_state_transitions.py: 422 lines (exceeded by 22)

## Solution
Extract shared test fixtures and helpers to conftest.py:
- Created tests/vibe3/orchestra/conftest.py (190 lines)
- Refactored both test files to use shared fixtures
- Removed duplicate helper functions

## Results
- test_dispatch_queue_operations.py: 310 lines (down from 438)
- test_dispatch_state_transitions.py: 314 lines (down from 422)
- All 141 orchestra tests pass
- Lint checks clean

## Verification
✅ All tests pass (15 tests in refactored files, 141 total in orchestra)
✅ LOC under 400-line CI block threshold
✅ ruff check clean
✅ No functionality lost
