+++ tests/vibe3/orchestra/conftest.py
@@ -1,0 +1,190 @@
+"""Shared fixtures and helpers for GlobalDispatchCoordinator tests."""
+
+from __future__ import annotations
+
+from unittest.mock import MagicMock
+
+import pytest
+
+from vibe3.models.orchestra_config import OrchestraConfig
+from vibe3.models.orchestration import IssueInfo, IssueState
+from vibe3.orchestra.global_dispatch_coordinator import (
+    GlobalDispatchCoordinator,
+)
+
+
+@pytest.fixture
+def make_issue() -> callable:
+    """Factory for creating mock issue objects."""
+
+    def _make_issue(number: int, priority: int = 5) -> MagicMock:
+        issue = MagicMock()
+        issue.number = number
+        issue.labels = [f"priority/{priority}"]
+        issue.milestone = None
+        issue.assignees = ["manager-bot"]
+        issue.priority = priority
+        return issue
+
+    return _make_issue
+
+
+@pytest.fixture
+def make_issue_info() -> callable:
+    """Factory for creating IssueInfo objects."""
+
+    def _make_issue_info(
+        number: int,
+        state: IssueState,
+        *,
+        assignees: list[str] | None = None,
+        labels: list[str] | None = None,
+    ) -> IssueInfo:
+        return IssueInfo(
+            number=number,
+            title=f"Issue {number}",
+            state=state,
+            labels=labels if labels is not None else [state.to_label()],
+            assignees=assignees if assignees is not None else ["manager-bot"],
+        )
+
+    return _make_issue_info
+
+
+@pytest.fixture
+def make_coordinator() -> callable:
+    """Factory for creating GlobalDispatchCoordinator with mocked dependencies."""
+
+    def _make_coordinator(
+        role: str = "manager",
+        ready_issues: list | None = None,
+        config: OrchestraConfig | None = None,
+        capacity: MagicMock | None = None,
+        with_branches: bool = False,
+    ) -> GlobalDispatchCoordinator:
+        if config is None:
+            config = OrchestraConfig(repo="owner/repo")
+            config.manager_usernames = ["manager-bot"]
+            config.supervisor_handoff.issue_label = "supervisor"
+
+        if capacity is None:
+            capacity = MagicMock()
+            capacity.config.max_concurrent_flows = 10
+            capacity.get_capacity_status = MagicMock(
+                return_value={
+                    "remaining": 10,
+                    "active_count": 0,
+                    "max_capacity": 10,
+                }
+            )
+            capacity._backend = None
+
+        github = MagicMock()
+
+        store = MagicMock()
+        store.db_path = ":memory:"
+        store.get_flow_state = MagicMock(return_value=None)
+        store.get_flows_by_issue = MagicMock(return_value=[])
+
+        flow_manager = MagicMock()
+        flow_manager.get_flow_for_issue = MagicMock(return_value=None)
+        flow_manager.git.branch_exists = MagicMock(return_value=True)
+
+        coordinator = GlobalDispatchCoordinator(
+            config=config,
+            capacity=capacity,
+            github=github,
+            store=store,
+            flow_manager=flow_manager,
+            registry=None,
+        )
+
+        if ready_issues:
+            role_map = {
+                "manager": IssueState.READY,
+                "handoff-manager": IssueState.HANDOFF,
+                "planner": IssueState.CLAIMED,
+                "plan": IssueState.CLAIMED,
+                "executor": IssueState.IN_PROGRESS,
+                "run": IssueState.IN_PROGRESS,
+                "reviewer": IssueState.REVIEW,
+                "review": IssueState.REVIEW,
+            }
+            target_state = role_map.get(role, IssueState.READY)
+
+            async def mock_poll(state: IssueState) -> list[IssueInfo]:
+                if state == target_state:
+                    issues_info = [
+                        IssueInfo(
+                            number=issue.number,
+                            title=f"Issue {issue.number}",
+                            state=target_state,
+                            labels=[
+                                target_state.to_label(),
+                                f"priority/{issue.priority}",
+                            ],
+                            assignees=["manager-bot"],
+                        )
+                        for issue in ready_issues
+                    ]
+                    return issues_info
+                return []
+
+            coordinator._poll_issues_by_state = mock_poll
+
+            if with_branches and role != "manager":
+
+                def mock_flow_context(issue_number: int) -> tuple[str, dict | None]:
+                    return (f"task/issue-{issue_number}", None)
+
+                coordinator._flow_context = mock_flow_context
+
+        return coordinator
+
+    return _make_coordinator
+
+
+@pytest.fixture
+def make_capacity() -> callable:
+    """Factory for creating mock capacity service."""
+
+    def _make_capacity(remaining: int = 1) -> MagicMock:
+        capacity = MagicMock()
+        capacity.config.max_concurrent_flows = max(remaining, 1)
+        capacity.get_capacity_status = MagicMock(
+            return_value={
+                "remaining": remaining,
+                "active_count": 0,
+                "max_capacity": max(remaining, 1),
+            }
+        )
+        capacity._run_command = MagicMock(
+            side_effect=Exception("tmux not available in tests")
+        )
+        capacity._backend = None
+        return capacity
+
+    return _make_capacity
+
+
+@pytest.fixture
+def install_issue_loader() -> callable:
+    """Factory for installing mock issue loader."""
+
+    def _install_issue_loader(
+        coordinator: GlobalDispatchCoordinator,
+        states: dict[int, IssueState | None],
+    ) -> None:
+        coordinator._load_issue = lambda issue_number: (
+            None
+            if states.get(issue_number) is None
+            else IssueInfo(
+                number=issue_number,
+                title=f"Issue {issue_number}",
+                state=states[issue_number],
+                labels=[states[issue_number].to_label()],
+                assignees=["manager-bot"],
+            )
+        )
+
+    return _install_issue_loader
