+++ tests/vibe3/orchestra/test_dispatch_state_transitions.py
@@ -3,160 +3,23 @@
 from __future__ import annotations
 
 import re
-from unittest.mock import MagicMock
 
 import pytest
 
-from vibe3.models.orchestra_config import OrchestraConfig
-from vibe3.models.orchestration import IssueInfo, IssueState
-from vibe3.orchestra.global_dispatch_coordinator import (
-    GlobalDispatchCoordinator,
-    QueueEntry,
-)
-
-
-def make_issue(number: int, priority: int = 5) -> MagicMock:
-    issue = MagicMock()
-    issue.number = number
-    issue.labels = [f"priority/{priority}"]
-    issue.milestone = None
-    issue.assignees = ["manager-bot"]
-    issue.priority = priority
-    return issue
-
-
-def make_issue_info(
-    number: int,
-    state: IssueState,
-    *,
-    assignees: list[str] | None = None,
-    labels: list[str] | None = None,
-) -> IssueInfo:
-    return IssueInfo(
-        number=number,
-        title=f"Issue {number}",
-        state=state,
-        labels=labels if labels is not None else [state.to_label()],
-        assignees=assignees if assignees is not None else ["manager-bot"],
-    )
-
-
-def make_coordinator(
-    role: str = "manager",
-    ready_issues: list | None = None,
-    config: OrchestraConfig | None = None,
-    capacity: MagicMock | None = None,
-    with_branches: bool = False,
-) -> GlobalDispatchCoordinator:
-    """Create a GlobalDispatchCoordinator with mocked dependencies."""
-    if config is None:
-        config = OrchestraConfig(repo="owner/repo")
-        config.manager_usernames = ["manager-bot"]
-        config.supervisor_handoff.issue_label = "supervisor"
-
-    if capacity is None:
-        capacity = MagicMock()
-        capacity.config.max_concurrent_flows = 10
-        capacity.get_capacity_status = MagicMock(
-            return_value={
-                "remaining": 10,
-                "active_count": 0,
-                "max_capacity": 10,
-            }
-        )
-        capacity._backend = None
-
-    github = MagicMock()
-
-    store = MagicMock()
-    store.db_path = ":memory:"
-    store.get_flow_state = MagicMock(return_value=None)
-    store.get_flows_by_issue = MagicMock(return_value=[])
-
-    flow_manager = MagicMock()
-    flow_manager.get_flow_for_issue = MagicMock(return_value=None)
-    flow_manager.git.branch_exists = MagicMock(return_value=True)
-
-    coordinator = GlobalDispatchCoordinator(
-        config=config,
-        capacity=capacity,
-        github=github,
-        store=store,
-        flow_manager=flow_manager,
-        registry=None,
-    )
-
-    if ready_issues:
-        role_map = {
-            "manager": IssueState.READY,
-            "handoff-manager": IssueState.HANDOFF,
-            "planner": IssueState.CLAIMED,
-            "plan": IssueState.CLAIMED,
-            "executor": IssueState.IN_PROGRESS,
-            "run": IssueState.IN_PROGRESS,
-            "reviewer": IssueState.REVIEW,
-            "review": IssueState.REVIEW,
-        }
-        target_state = role_map.get(role, IssueState.READY)
-
-        async def mock_poll(state: IssueState) -> list[IssueInfo]:
-            if state == target_state:
-                issues_info = [
-                    IssueInfo(
-                        number=issue.number,
-                        title=f"Issue {issue.number}",
-                        state=target_state,
-                        labels=[target_state.to_label(), f"priority/{issue.priority}"],
-                        assignees=["manager-bot"],
-                    )
-                    for issue in ready_issues
-                ]
-                return issues_info
-            return []
-
-        coordinator._poll_issues_by_state = mock_poll
-
-        if with_branches and role != "manager":
-
-            def mock_flow_context(issue_number: int) -> tuple[str, dict | None]:
-                return (f"task/issue-{issue_number}", None)
-
-            coordinator._flow_context = mock_flow_context
-
-    return coordinator
-
-
-def make_capacity(remaining: int = 1) -> MagicMock:
-    capacity = MagicMock()
-    capacity.config.max_concurrent_flows = max(remaining, 1)
-    capacity.get_capacity_status = MagicMock(
-        return_value={
-            "remaining": remaining,
-            "active_count": 0,
-            "max_capacity": max(remaining, 1),
-        }
-    )
-    capacity._run_command = MagicMock(
-        side_effect=Exception("tmux not available in tests")
-    )
-    capacity._backend = None
-    return capacity
-
-
-def install_issue_loader(
-    coordinator: GlobalDispatchCoordinator,
-    states: dict[int, IssueState | None],
-) -> None:
-    coordinator._load_issue = lambda issue_number: (
-        None
-        if states.get(issue_number) is None
-        else make_issue_info(issue_number, states[issue_number])
-    )
+from vibe3.models.orchestration import IssueState
+from vibe3.orchestra.global_dispatch_coordinator import QueueEntry
 
 
 class TestStateTransitions:
     @pytest.mark.asyncio
-    async def test_collect_order_prefers_higher_state_roles_first(self) -> None:
+    async def test_collect_order_prefers_higher_state_roles_first(
+        self,
+        make_issue,
+        make_capacity,
+        make_coordinator,
+        install_issue_loader,
+        make_issue_info,
+    ) -> None:
         """Test that issues are collected in priority order."""
         _ = make_issue(1)
         _ = make_issue(2)
@@ -166,7 +29,7 @@ class TestStateTransitions:
             "planner", [], capacity=capacity, with_branches=True
         )
 
-        async def mock_poll(state: IssueState) -> list[IssueInfo]:
+        async def mock_poll(state: IssueState) -> list:
             if state == IssueState.READY:
                 return [make_issue_info(1, IssueState.READY)]
             elif state == IssueState.CLAIMED:
@@ -194,7 +57,14 @@ class TestStateTransitions:
         assert emit_calls[0][1].number == 2
 
     @pytest.mark.asyncio
-    async def test_capacity_limit_stops_dispatch(self) -> None:
+    async def test_capacity_limit_stops_dispatch(
+        self,
+        make_issue,
+        make_capacity,
+        make_coordinator,
+        install_issue_loader,
+        make_issue_info,
+    ) -> None:
         """Test that capacity limit stops dispatch after limit reached."""
         _ = make_issue(304)
         _ = make_issue(303)
@@ -205,7 +75,7 @@ class TestStateTransitions:
             "planner", [], capacity=capacity, with_branches=True
         )
 
-        async def mock_poll(state: IssueState) -> list[IssueInfo]:
+        async def mock_poll(state: IssueState) -> list:
             if state == IssueState.REVIEW:
                 return [make_issue_info(304, IssueState.REVIEW)]
             elif state == IssueState.CLAIMED:
@@ -239,7 +109,14 @@ class TestStateTransitions:
         assert 372 not in dispatched_numbers
 
     @pytest.mark.asyncio
-    async def test_state_change_requeues_issue_to_front(self) -> None:
+    async def test_state_change_requeues_issue_to_front(
+        self,
+        make_issue,
+        make_capacity,
+        make_coordinator,
+        install_issue_loader,
+        make_issue_info,
+    ) -> None:
         """Test that issue with state change gets promoted to front of queue."""
         _ = make_issue(1)
         _ = make_issue(2)
@@ -249,7 +126,7 @@ class TestStateTransitions:
             "planner", [], capacity=capacity, with_branches=True
         )
 
-        async def mock_poll(state: IssueState) -> list[IssueInfo]:
+        async def mock_poll(state: IssueState) -> list:
             if state == IssueState.READY:
                 return [
                     make_issue_info(1, IssueState.READY),
@@ -276,7 +153,7 @@ class TestStateTransitions:
 
         states[1] = IssueState.CLAIMED
 
-        async def mock_poll_claimed(state: IssueState) -> list[IssueInfo]:
+        async def mock_poll_claimed(state: IssueState) -> list:
             if state == IssueState.CLAIMED:
                 return [make_issue_info(1, IssueState.CLAIMED)]
             return []
@@ -289,14 +166,21 @@ class TestStateTransitions:
         assert emit_calls[1][1].number == 1
 
     @pytest.mark.asyncio
-    async def test_terminal_state_removes_issue_from_queue(self) -> None:
+    async def test_terminal_state_removes_issue_from_queue(
+        self,
+        make_issue,
+        make_capacity,
+        make_coordinator,
+        install_issue_loader,
+        make_issue_info,
+    ) -> None:
         """Test that BLOCKED issues are removed from queue."""
         manager_issue = make_issue(1)
         capacity = make_capacity(remaining=1)
 
         coordinator = make_coordinator("manager", [manager_issue], capacity=capacity)
 
-        async def mock_poll(state: IssueState) -> list[IssueInfo]:
+        async def mock_poll(state: IssueState) -> list:
             if state == IssueState.READY:
                 return [make_issue_info(1, IssueState.READY)]
             return []
@@ -330,6 +214,10 @@ class TestLoggingBehavior:
     @pytest.mark.asyncio
     async def test_logs_dispatch_intent_instead_of_dispatch_success(
         self,
+        make_issue,
+        make_capacity,
+        make_coordinator,
+        install_issue_loader,
         monkeypatch: pytest.MonkeyPatch,
     ) -> None:
         planner_issue = make_issue(303)
@@ -373,6 +261,10 @@ class TestLoggingBehavior:
     @pytest.mark.asyncio
     async def test_logs_dispatch_intent_before_emit_side_effect(
         self,
+        make_issue,
+        make_capacity,
+        make_coordinator,
+        install_issue_loader,
         monkeypatch: pytest.MonkeyPatch,
     ) -> None:
         planner_issue = make_issue(467)
